Fix signed vs unsigned comparison issue
C89 rules are that in mixed signed / unsigned comparisons, if the
operands are smaller than int, they are promited to int. This is what
occurs on modern systems, and it is the reason this code works fine
there.

If the operands have size greater or equal to int and have the same size
as each other, they are promoted to the unsigned type. This causes
negative values to be treated as very large positive values. On the
Plexus, sizeof(int) == sizeof(int16_t) == sizeof(uint16_t).

Whenever delta was < 0, the piece would teleport all the way to the left
side of the board.

+++ src/main.c
@@ -1086,7 +1086,7 @@ play_game(uint16_t initial_level, uint16_t seed, bool two_player)
 		rotation = (rotation + dir) & 3;
 
 		delta = piece->f[old_rotation].shift - piece->f[rotation].shift;
-		x = x >= delta ? x - delta : 0;
+		x = (int16_t)x >= delta ? x - delta : 0;
 		break;
 	    }
